+++ src/components/CountryMap.js
@@ -1,0 +1,34 @@
+import React from 'react' // eslint-disable-line no-unused-vars
+
+
+class CountryMap extends React.Component {
+  constructor (props) {
+    super(props)
+    this.state = {'SVG': ''}
+  }
+  componentWillMount () {
+    let component = this
+    // TODO move to country-specific bucket or at least folder
+    fetch('https://s3.amazonaws.com/assets.participedia.xyz/' + this.props.countrycode + '.svg').then(function (response) {
+      return response.text()
+    }).then(function (SVGtext) {
+      let svg = '<svg width="100%" height="100%" xmlns="http://www.w3.org/2000/svg" version="1.1"><defs><style type="text/css"><![CDATA[path {stroke: none;fill: #ff6f00;}]]></style></defs>' + SVGtext + '</svg>'
+      component.setState({SVG: svg})
+    })
+  }
+
+  render () {
+    return ( 
+      <div>
+        <div dangerouslySetInnerHTML={{__html: this.state.SVG}} />
+        { this.props.city? 
+          <p className="case-location">{this.props.city}, {this.props.countrycode}</p>
+          :
+          <p className="case-location">{this.props.countrycode}</p>
+        }
+      </div>
+    )
+  }
+}
+
+export default CountryMap
