+++ src/components/SearchHit.scss
@@ -63,7 +63,6 @@
 
 .thumbnail {
   width: 100%;
-  // width: 288px;
   height: 160px;
   background-size: cover;
   background-position: center;
